Move PythonCodeEditor to separate component
