+++ enterprise/frontend/src/metabase-enterprise/transforms/components/PythonEditor/PythonEditor.module.css
@@ -1,0 +1,4 @@
+.editor {
+  width: 100%;
+  height: 100%;
+}
