+++ enterprise/frontend/src/metabase-enterprise/transforms/components/PythonEditor/index.ts
@@ -1,0 +1,1 @@
+export * from "./PythonEditor";
